+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.0.1
+
+- add ext_context param for external file with context
+- allow separate templates to fail, the preprocessor would issue warning and skip them
+
 # 1.0.0
 
 -    Initial release